+++ src/openApi/v3/interfaces/ModelConfig.d.ts
@@ -1,3 +1,4 @@
+import { GetTypeName } from '../parser/getType';
 import { OpenApi } from './OpenApi';
 import { OpenApiSchema } from './OpenApiSchema';
 
@@ -7,4 +8,6 @@ export interface ModelConfig {
     isDefinition?: boolean;
     name?: string;
     path?: string;
+    getTypeByRef: GetTypeName;
+    parentRef: string;
 }