+++ src/openApi/v3/interfaces/OpenApiRequestBody.d.ts
@@ -1,6 +1,6 @@
 import type { Dictionary } from '../../../utils/types';
-import type { OpenApiMediaType } from './OpenApiMediaType';
 import type { OpenApiReference } from '../../interfaces/OpenApiReference';
+import type { OpenApiMediaType } from './OpenApiMediaType';
 
 /**
  * https://github.com/OAI/OpenAPI-Specification/blob/master/versions/3.0.2.md#requestBodyObject