+++ src/openApi/v3/parser/stripNamespace.ts
@@ -9,13 +9,16 @@ import { hasMappedType } from './getMappedType';
  * @param value
  */
 export function stripNamespace(value: string): string {
+    if (!value || hasMappedType(value)) {
+        return value;
+    }
     if (!value.match(/^(http:\/\/|https:\/\/|#\/)/g) && !hasMappedType(value) && !value.match(/^array\[[a-z]+\]$/g)) {
         const directoryName = dirName(value);
         const extName = extname(value);
         const baseName = extName.toLowerCase().match(/(\.json|\.yaml|\.yml)$/g) ? getClassName(basename(value, extName)) : getClassName(basename(value));
         return directoryName ? join(directoryName, baseName) : baseName;
     }
-    return value
+    const clearValue = value
         .trim()
         .replace(/^#\/components\/schemas\//, '')
         .replace(/^#\/components\/responses\//, '')
@@ -26,4 +29,7 @@ export function stripNamespace(value: string): string {
         .replace(/^#\/components\/securitySchemes\//, '')
         .replace(/^#\/components\/links\//, '')
         .replace(/^#\/components\/callbacks\//, '');
+    const directoryName = dirName(clearValue);
+    const baseName = getClassName(basename(clearValue));
+    return directoryName ? join(directoryName, baseName) : baseName;
 }