+++ src/utils/getAbsolutePath.ts
@@ -1,0 +1,16 @@
+import { dirName, join } from '../core/path';
+
+export function getAbsolutePath(definitionRef: string | undefined, parentRef: string): string {
+    if (definitionRef) {
+        if (parentRef && parentRef.match(/^(#\/)/g)) {
+            return definitionRef;
+        } else if (definitionRef.match(/^(#\/)/g)) {
+            return `${parentRef}${definitionRef}`;
+        }
+        return join(dirName(parentRef), definitionRef);
+    }
+    if (parentRef.match(/^(#\/)/g)) {
+        return '';
+    }
+    return parentRef;
+}