+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.30] - 2021-11-20
+
+### Added
+
+- Add prefix to class name (I - interface, E - enum, T - type)
+
+### Fixed
+
+- Starts a class name from a component block with capital char
+
 ## [0.1.29] - 2021-11-19
 
 ### Added